refactor: move ShellEditor to new file
